astro-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyswisseph>=2.10.3.2
pytz>=2023.3
timezonefinder>=6.2.0
geopy>=2.4.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
